Update tests to nimut, move all required fixtures to Tests\Unit\
